Refactor and add test for make_post_header_replacements
