+++ src/tests/fixtures/ewhc-ch-2023-257_replacements.txt
@@ -1,0 +1,72 @@
+{"case": ["[2022] UKSC 3", "[2022] UKSC 3", "2022", "https://caselaw.nationalarchives.gov.uk/uksc/2022/3", true]}
+{"case": ["[1975] AC 591", "[1975] AC 591", "1975", "#", false]}
+{"case": ["[2001] AC 349", "[2001] AC 349", "2001", "#", false]}
+{"case": ["[1993] AC 593", "[1993] AC 593", "1993", "#", false]}
+{"case": ["[1981] AC 251", "[1981] AC 251", "1981", "#", false]}
+{"case": ["[1993] AC 593", "[1993] AC 593", "1993", "#", false]}
+{"case": ["[2022] UKSC 3", "[2022] UKSC 3", "2022", "https://caselaw.nationalarchives.gov.uk/uksc/2022/3", true]}
+{"case": ["[2022] 2 WLR 343", "[2022] 2 WLR 343", "2022", "#", false]}
+{"case": ["[1980] 1 Ch 576", "[1980] 1 Ch 576", "1980", "#", false]}
+{"case": ["[1987] 1 All ER 528", "[1987] 1 All ER 528", "1987", "#", false]}
+{"case": ["[2005] EWCA Civ 1004", "[2005] EWCA Civ 1004", "2005", "https://caselaw.nationalarchives.gov.uk/ewca/civ/2005/1004", true]}
+{"case": ["[2014] EWHC 1178 (Ch)", "[2014] EWHC 1178 (Ch)", "2014", "https://caselaw.nationalarchives.gov.uk/ewhc/ch/2014/1178", true]}
+{"case": ["[2008] EWHC 153 (Ch)", "[2008] EWHC 153 (Ch)", "2008", "https://caselaw.nationalarchives.gov.uk/ewhc/ch/2008/153", true]}
+{"case": ["[1937] 1 KB 209", "[1937] 1 KB 209", "1937", "#", false]}
+{"leg": ["Coronavirus Act 2020", "http://www.legislation.gov.uk/id/ukpga/2020/7", "2020 c. 7"]}
+{"leg": ["Corporation Tax Act 2010", "http://www.legislation.gov.uk/id/ukpga/2010/4", "2010 c. 4"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["FA 2014", "http://www.legislation.gov.uk/id/ukpga/2014/26", "2014 c. 26"]}
+{"leg": ["FA 2014", "http://www.legislation.gov.uk/id/ukpga/2014/26", "2014 c. 26"]}
+{"leg": ["FA 2014", "http://www.legislation.gov.uk/id/ukpga/2014/26", "2014 c. 26"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Finance Act 2004", "http://www.legislation.gov.uk/id/ukpga/2004/12", "2004 c. 12"]}
+{"leg": ["Insolvency Act 1986", "http://www.legislation.gov.uk/id/ukpga/1986/45", "1986 c. 45"]}
+{"leg": ["Reserve Forces Act 1980", "http://www.legislation.gov.uk/id/ukpga/1980/9", "1980 c. 9"]}
+{"leg": ["Reserve Forces Act 1996", "http://www.legislation.gov.uk/id/ukpga/1996/14", "1996 c. 14"]}
+{"leg": ["Social Security Act 1975", "http://www.legislation.gov.uk/id/ukpga/1975/14", "1975 c. 14"]}
